+++ shopify/packages/ajax-cart-api/src/effect/api/CartAdd.ts
@@ -1,12 +1,11 @@
 import { CartAddInput, CartAddOutput } from "../schema.js";
 
-import * as AjaxRequest from "../services/AjaxRequest.js";
+import * as AjaxClient from "@brytdesigns/shopify-ajax-client";
 
 export type AddInput = CartAddInput;
 
-export const make = AjaxRequest.makeFactory({
+export const make = AjaxClient.makeFactory({
   routeName: "cart_add_url",
-  method: "post",
   inputSchema: CartAddInput,
   outputSchema: CartAddOutput,
 });