+++ shopify/packages/ajax-cart-api/src/effect/api/CartChange.ts
@@ -1,12 +1,11 @@
 import { CartChangeInput, CartChangeOutput } from "../schema.js";
 
-import * as AjaxRequest from "../services/AjaxRequest.js";
+import * as AjaxClient from "@brytdesigns/shopify-ajax-client";
 
 export type ChangeInput = CartChangeInput;
 
-export const make = AjaxRequest.makeFactory({
+export const make = AjaxClient.makeFactory({
   routeName: "cart_change_url",
-  method: "post",
   inputSchema: CartChangeInput,
   outputSchema: CartChangeOutput,
 });