+++ shopify/packages/ajax-cart-api/src/effect/api/CartUpdate.ts
@@ -1,12 +1,11 @@
 import { CartUpdateInput, CartUpdateOutput } from "../schema.js";
 
-import * as AjaxRequest from "../services/AjaxRequest.js";
+import * as AjaxClient from "@brytdesigns/shopify-ajax-client";
 
 export type UpdateInput = CartUpdateInput;
 
-export const make = AjaxRequest.makeFactory({
+export const make = AjaxClient.makeFactory({
   routeName: "cart_update_url",
-  method: "post",
   inputSchema: CartUpdateInput,
   outputSchema: CartUpdateOutput,
 });