+++ shopify/packages/ajax-cart-api/src/index.ts
@@ -3,9 +3,8 @@ import * as Logger from "effect/Logger";
 import * as Layer from "effect/Layer";
 import * as LogLevel from "effect/LogLevel";
 import * as API from "./effect/index.js";
-import * as AjaxRequest from "./effect/services/AjaxRequest.js";
 import * as Schema from "effect/Schema";
-import { AjaxClientResponse } from "./effect/data/index.js";
+import * as AjaxClient from "@brytdesigns/shopify-ajax-client";
 import { CartOutput } from "./effect/schema.js";
 
 export namespace createAjaxCartClient {
@@ -35,13 +34,7 @@ export const createAjaxCartApi = ({ debug = false }) => {
     minimumLogLevel = LogLevel.All;
   }
 
-  // if (logger) {
-  //   loggerLayer = Logger.replace(Logger.defaultLogger, Logger.make(logger));
-  //   baseLayer = Layer.mergeAll(loggerLayer);
-  //   minimumLogLevel = LogLevel.All;
-  // }
-
-  const ajaxLayer = Layer.mergeAll(baseLayer, AjaxRequest.Default);
+  const ajaxLayer = Layer.mergeAll(baseLayer, AjaxClient.Default);
 
   return {
     add: (
@@ -49,38 +42,15 @@ export const createAjaxCartApi = ({ debug = false }) => {
       options?: createAjaxCartClient.RequestOptions,
     ) =>
       Effect.runPromise(
-        API.add(input, options).pipe(
+        API.add(input, {
+          ...(options || {}),
+          headers: {
+            "X-SK-Variant-Source": "shopify-ajax-cart-api",
+            ...(options?.headers || {}),
+          },
+        }).pipe(
           Logger.withMinimumLogLevel(minimumLogLevel),
           Effect.provide(ajaxLayer),
-          Effect.catchAll((error) => {
-            if (
-              error._tag === "ParseError" ||
-              error._tag === "RequestError" ||
-              error._tag === "ResponseError" ||
-              error._tag ===
-                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
-            ) {
-              return Effect.fail(
-                new Error(error.message, {
-                  cause: error.cause,
-                }),
-              );
-            }
-            if (error._tag === "HttpBodyError") {
-              const reason = error.reason;
-              if (reason._tag === "JsonError") {
-                return Effect.fail(new Error("Failed to parse JSON response"));
-              }
-              if (reason._tag === "SchemaError") {
-                return Effect.fail(
-                  new Error(reason.error.message, {
-                    cause: reason.error.cause,
-                  }),
-                );
-              }
-            }
-            return Effect.fail(new Error(error.toString()));
-          }),
         ),
         {
           signal: options?.signal,
@@ -91,38 +61,15 @@ export const createAjaxCartApi = ({ debug = false }) => {
       options?: createAjaxCartClient.RequestOptions,
     ) =>
       Effect.runPromise(
-        API.change(input, options).pipe(
+        API.change(input, {
+          ...(options || {}),
+          headers: {
+            "X-SK-Variant-Source": "shopify-ajax-cart-api",
+            ...(options?.headers || {}),
+          },
+        }).pipe(
           Logger.withMinimumLogLevel(minimumLogLevel),
           Effect.provide(ajaxLayer),
-          Effect.catchAll((error) => {
-            if (
-              error._tag === "ParseError" ||
-              error._tag === "RequestError" ||
-              error._tag === "ResponseError" ||
-              error._tag ===
-                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
-            ) {
-              return Effect.fail(
-                new Error(error.message, {
-                  cause: error.cause,
-                }),
-              );
-            }
-            if (error._tag === "HttpBodyError") {
-              const reason = error.reason;
-              if (reason._tag === "JsonError") {
-                return Effect.fail(new Error("Failed to parse JSON response"));
-              }
-              if (reason._tag === "SchemaError") {
-                return Effect.fail(
-                  new Error(reason.error.message, {
-                    cause: reason.error.cause,
-                  }),
-                );
-              }
-            }
-            return Effect.fail(new Error(error.toString()));
-          }),
         ),
         {
           signal: options?.signal,
@@ -133,81 +80,34 @@ export const createAjaxCartApi = ({ debug = false }) => {
       options?: createAjaxCartClient.RequestOptions,
     ) =>
       Effect.runPromise(
-        API.clear(input, options).pipe(
+        API.clear(input, {
+          ...(options || {}),
+          headers: {
+            "X-SK-Variant-Source": "shopify-ajax-cart-api",
+            ...(options?.headers || {}),
+          },
+        }).pipe(
           Logger.withMinimumLogLevel(minimumLogLevel),
           Effect.provide(ajaxLayer),
-          Effect.catchAll((error) => {
-            if (
-              error._tag === "ParseError" ||
-              error._tag === "RequestError" ||
-              error._tag === "ResponseError" ||
-              error._tag ===
-                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
-            ) {
-              return Effect.fail(
-                new Error(error.message, {
-                  cause: error.cause,
-                }),
-              );
-            }
-            if (error._tag === "HttpBodyError") {
-              const reason = error.reason;
-              if (reason._tag === "JsonError") {
-                return Effect.fail(new Error("Failed to parse JSON response"));
-              }
-              if (reason._tag === "SchemaError") {
-                return Effect.fail(
-                  new Error(reason.error.message, {
-                    cause: reason.error.cause,
-                  }),
-                );
-              }
-            }
-            return Effect.fail(new Error(error.toString()));
-          }),
         ),
         {
           signal: options?.signal,
         },
       ),
-
     get: (
       input: API.get.Input,
       options?: createAjaxCartClient.RequestOptions,
     ) =>
       Effect.runPromise(
-        API.get(input, options).pipe(
+        API.get(input, {
+          ...(options || {}),
+          headers: {
+            "X-SK-Variant-Source": "shopify-ajax-cart-api",
+            ...(options?.headers || {}),
+          },
+        }).pipe(
           Logger.withMinimumLogLevel(minimumLogLevel),
           Effect.provide(ajaxLayer),
-          Effect.catchAll((error) => {
-            if (
-              error._tag === "ParseError" ||
-              error._tag === "RequestError" ||
-              error._tag === "ResponseError" ||
-              error._tag ===
-                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
-            ) {
-              return Effect.fail(
-                new Error(error.message, {
-                  cause: error.cause,
-                }),
-              );
-            }
-            if (error._tag === "HttpBodyError") {
-              const reason = error.reason;
-              if (reason._tag === "JsonError") {
-                return Effect.fail(new Error("Failed to parse JSON response"));
-              }
-              if (reason._tag === "SchemaError") {
-                return Effect.fail(
-                  new Error(reason.error.message, {
-                    cause: reason.error.cause,
-                  }),
-                );
-              }
-            }
-            return Effect.fail(new Error(error.toString()));
-          }),
         ),
         {
           signal: options?.signal,
@@ -219,42 +119,15 @@ export const createAjaxCartApi = ({ debug = false }) => {
       options?: createAjaxCartClient.RequestOptions,
     ) =>
       Effect.runPromise(
-        API.update(input, options).pipe(
+        API.update(input, {
+          ...(options || {}),
+          headers: {
+            "X-SK-Variant-Source": "shopify-ajax-cart-api",
+            ...(options?.headers || {}),
+          },
+        }).pipe(
           Logger.withMinimumLogLevel(minimumLogLevel),
           Effect.provide(ajaxLayer),
-          Effect.catchAll((error) => {
-            if (
-              error._tag === "ParseError" ||
-              error._tag === "RequestError" ||
-              error._tag === "ResponseError" ||
-              error._tag ===
-                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
-            ) {
-              return Effect.succeed(
-                AjaxClientResponse.make({
-                  error: {
-                    status: 400,
-                    message: error.message,
-                    description: "Something went wrong",
-                  },
-                }),
-              );
-            }
-            if (error._tag === "HttpBodyError") {
-              const reason = error.reason;
-              if (reason._tag === "JsonError") {
-                return Effect.fail(new Error("Failed to parse JSON response"));
-              }
-              if (reason._tag === "SchemaError") {
-                return Effect.fail(
-                  new Error(reason.error.message, {
-                    cause: reason.error.cause,
-                  }),
-                );
-              }
-            }
-            return Effect.fail(new Error(error.toString()));
-          }),
         ),
         {
           signal: options?.signal,