+++ shopify/packages/ajax-client/src/LoggerUtils.ts
@@ -1,7 +1,7 @@
 import { Logging } from "@brytdesigns/shopify-utils/effect";
 
 export const withNamespacedLogSpan = Logging.makeNamespacedLogSpan(
-  "[@brytdesigns/shopify-ajax-cart-client]",
+  "[@brytdesigns/shopify-ajax-client]",
 );
 
 export const filterLevelOrNever = Logging.filterLevelOrNever;